First, rework the deltas thing a bit, define colors using a color: member instead, this lets us have a different look for the delta line and marker depending if the datapoint is above or below the delta marker.
Also, fix drawing of points on the far right side of the graph by pushing the graph in 10 pixels (probably should be less) in the div. As a result we can also stop chopping the marker drawing to the size of the plot area (since we can draw over the edge just like points do).